refactor(core-kvalue): move settings to dedicated module.
