+++ features/schedules/schedules-sample/src/androidTest/kotlin/com/paligot/confily/schedules/sample/FilteringScheduleTest.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.test.core.app.ActivityScenario
 import androidx.test.platform.app.InstrumentationRegistry
 import com.paligot.confily.core.agenda.AgendaDao
-import com.paligot.confily.core.db.ConferenceSettings
+import com.paligot.confily.core.kvalue.ConferenceSettings
 import com.paligot.confily.core.sample.BuildConfig
 import com.paligot.confily.core.schedules.ScheduleDao
 import com.paligot.confily.core.test.instrumentedModule
