+++ settings.gradle.kts
@@ -69,6 +69,7 @@ include(":shared:core-api")
 include(":shared:core-db")
 include(":shared:core-di")
 include(":shared:core-fs")
+include(":shared:core-kvalue")
 include(":shared:models")
 include(":shared:resources")
 include(":shared:ui-models")
