+++ shared/core-di/build.gradle.kts
@@ -27,6 +27,7 @@ kotlin {
                 export(projects.shared.coreApi)
                 export(projects.shared.coreDb)
                 export(projects.shared.core)
+                export(projects.shared.coreKvalue)
                 export(projects.shared.models)
                 export(projects.shared.uiModels)
                 export(projects.shared.resources)
@@ -42,6 +43,7 @@ kotlin {
                 api(projects.shared.core)
                 api(projects.shared.coreApi)
                 api(projects.shared.coreDb)
+                api(projects.shared.coreKvalue)
                 api(projects.shared.resources)
                 implementation(libs.koin.core)
                 implementation(libs.ktor.client.core)
