+++ shared/core/build.gradle.kts
@@ -28,6 +28,7 @@ kotlin {
                 export(projects.shared.coreApi)
                 export(projects.shared.coreDb)
                 export(projects.shared.coreFs)
+                export(projects.shared.coreKvalue)
                 export(projects.shared.models)
                 export(projects.shared.uiModels)
                 export(projects.shared.resources)
@@ -43,6 +44,7 @@ kotlin {
                 api(projects.shared.coreApi)
                 api(projects.shared.coreDb)
                 api(projects.shared.coreFs)
+                api(projects.shared.coreKvalue)
                 api(projects.shared.models)
                 api(projects.shared.uiModels)
                 api(projects.shared.resources)
