+++ shared/core/src/commonMain/kotlin/com/paligot/confily/core/networking/NetworkingRepository.kt
@@ -1,7 +1,7 @@
 package com.paligot.confily.core.networking
 
 import com.paligot.confily.core.QrCodeGenerator
-import com.paligot.confily.core.db.ConferenceSettings
+import com.paligot.confily.core.kvalue.ConferenceSettings
 import com.paligot.confily.core.extensions.encodeToString
 import com.paligot.confily.core.fs.ConferenceFileSystem
 import com.paligot.confily.models.ui.ExportNetworkingUi
